sport entity 위치 변경
