+++ src/main/java/com/beyond/sportsmatch/domain/match/model/entity/MatchApplication.java
@@ -1,6 +1,7 @@
 package com.beyond.sportsmatch.domain.match.model.entity;
 
 import com.beyond.sportsmatch.domain.match.model.dto.MatchRequestDto;
+import com.beyond.sportsmatch.domain.user.model.entity.Sport;
 import com.beyond.sportsmatch.domain.user.model.entity.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
