--- src/main/java/com/beyond/sportsmatch/domain/match/model/entity/Sport.java
@@ -1,23 +1,0 @@
-package com.beyond.sportsmatch.domain.match.model.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Data;
-
-@Entity
-@Data
-public class Sport {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
-    @Column
-    private String name;
-
-    @Column
-    private int requiredPersonnel;
-}
