+++ src/main/java/com/beyond/sportsmatch/domain/user/model/entity/Sport.java
@@ -1,19 +1,23 @@
-//package com.beyond.sportsmatch.domain.user.model.entity;
-//
-//import jakarta.persistence.*;
-//import lombok.Getter;
-//import lombok.Setter;
-//
-//@Entity
-//@Getter
-//@Setter
-//@Table(name = "sport")
-//public class Sport {
-//
-//    @Id
-//    @GeneratedValue(strategy = GenerationType.IDENTITY)
-//    private Integer id;       // PK
-//
-//    @Column(nullable = false, unique = true, length = 100)
-//    private String name;      // 운동 이름
-//}
+package com.beyond.sportsmatch.domain.user.model.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Data;
+
+@Entity
+@Data
+public class Sport {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @Column
+    private String name;
+
+    @Column
+    private int requiredPersonnel;
+}
