+++ src/main/java/com/beyond/sportsmatch/domain/user/model/repository/SportRepository.java
@@ -2,7 +2,7 @@ package com.beyond.sportsmatch.domain.user.model.repository;
 
 
 
-import com.beyond.sportsmatch.domain.match.model.entity.Sport;
+import com.beyond.sportsmatch.domain.user.model.entity.Sport;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
